# frontend/src/renderer/src/types/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "BackupRequest": {
        "description": "Request body for the explicit conversation backup command.",
        "properties": {
          "conversation_ids": {
            "anyOf": [
              {
                "items": {
                  "format": "uuid",
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Ids"
          },
          "dest_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dest Dir"
          }
        },
        "title": "BackupRequest",
        "type": "object"
      },
      "BackupResult": {
        "description": "Outcome of an explicit conversation backup.",
        "properties": {
          "exported": {
            "title": "Exported",
            "type": "integer"
          },
          "path": {
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "exported",
          "path"
        ],
        "title": "BackupResult",
        "type": "object"
      },
      "Body_upload_image_api_chat_upload_post": {
        "properties": {
          "conversation_id": {
            "description": "Target conversation UUID",
            "title": "Conversation Id",
            "type": "string"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "Image file (jpg/png/gif/webp)",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "file"
        ],
        "title": "Body_upload_image_api_chat_upload_post",
        "type": "object"
      },
...
      "ConversationExport": {
        "description": "Full conversation export (REST response body and backup file schema).",
        "properties": {
          "active_versions": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Active Versions",
            "type": "object"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ExportedMessage"
            },
            "title": "Messages",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "created_at",
          "updated_at",
          "messages"
        ],
        "title": "ConversationExport",
        "type": "object"
      },
...
      "ExportedAttachment": {
        "description": "One attachment inside an exported message.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "file_id": {
            "title": "File Id",
            "type": "string"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "file_id",
          "url",
          "filename",
          "content_type",
          "file_path"
        ],
        "title": "ExportedAttachment",
        "type": "object"
      },
      "ExportedMessage": {
        "description": "One message inside a conversation export.",
        "properties": {
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ExportedAttachment"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "context_excluded": {
            "default": false,
            "title": "Context Excluded",
            "type": "boolean"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_context_summary": {
            "default": false,
            "title": "Is Context Summary",
            "type": "boolean"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "thinking_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thinking Content"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          },
          "tool_calls": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Calls"
          },
          "version_group_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Group Id"
          },
          "version_index": {
            "default": 0,
            "title": "Version Index",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "role",
          "content",
          "created_at"
        ],
        "title": "ExportedMessage",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AL\\CE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/chat/conversations/backup": {
      "post": {
        "description": "Export conversations as JSON files to an explicit destination.\n\nThis is the user-facing replacement of the removed automatic JSON\nmirror (spec §5.2): SQLite is the single source of truth and backups\nhappen only on explicit command (UI entry or agent tool).",
        "operationId": "backup_conversations_api_chat_conversations_backup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backup Conversations",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/chat/conversations/{conversation_id}/export": {
      "get": {
        "description": "Export a full conversation with all messages and metadata.\n\nArgs:\n    conversation_id: UUID of the conversation to export.\n\nReturns:\n    The complete conversation JSON including messages and attachments.",
        "operationId": "export_conversation_api_chat_conversations__conversation_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationExport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Conversation",
        "tags": [
          "chat"
        ]
      }
    },
...
    "/api/chat/upload": {
      "post": {
        "description": "Upload an image for use with vision-capable models.\n\nSaves the file to ``data/uploads/{conversation_id}/`` and creates a\npending :class:`Attachment` record (``message_id`` is set later when the\nWebSocket message referencing this file is sent).\n\nReturns:\n    A dict with ``file_id``, ``url``, ``filename``, and ``content_type``.\n\nRaises:\n    HTTPException 400: If the file type is not an allowed image format.\n    HTTPException 413: If the file exceeds the configured size limit.",
        "operationId": "upload_image_api_chat_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_image_api_chat_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Image Api Chat Upload Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Image",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}